restruture the repository
- remove extra dockfiles
- remove debug files
- add usage info to the readme
- add help to scripts
- fix bugs in the main dockerfile

+++ clean_docker.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
+if [ "$1" = '-h' ]; then
+    echo 'usage: $0'
+    echo 'removes all containers and removes exited containers'
+    exit 0
+fi
+
 docker stop $(sudo docker ps -a -q)
 docker rm $(sudo docker ps -a -q)
